os kill in only applied in case of ROOT genfit

+++ python/shipRoot_conf.py
@@ -1,10 +1,13 @@
 from __future__ import print_function
-import ROOT, atexit, sys, os
+import ROOT
+import atexit
+import sys
+import os
 from pythia8_conf_utils import addHNLtoROOT
 from pythia8darkphoton_conf import addDPtoROOT
 
 # Try to check if config has been executed...
-if os.environ.get('FAIRSHIP_ROOT','') == '' and os.environ.get('Linux_Flavour_','') == '' :
+if os.environ.get('FAIRSHIP_ROOT', '') == '' and os.environ.get('Linux_Flavour_', '') == '':
    print("Do first: source config.[c]sh")
    quit()
 
@@ -40,18 +43,24 @@ else:
     ROOT.gSystem.Load("libG4clhep")
 ROOT.gInterpreter.ProcessLine('typedef double Double32_t')
 
-#-----prepare python exit-----------------------------------------------
+# -----prepare python exit-----------------------------------------------
+
+
 def pyExit():
-   x = sys.modules['__main__']
-   if hasattr(x,'run'): 
-        del x.run
-        print("make suicid, until better solution found to ROOT/genfit interference")
-        for f in ROOT.gROOT.GetListOfFiles(): 
-         if f.IsWritable() and f.IsOpen(): f.Close()
-        os.system('kill '+str(os.getpid()))
-   if hasattr(x,'fMan'): del x.fMan
-   if hasattr(x,'fRun'): del x.fRun
-   pass
+    for module in sys.modules:
+        if "ROOT.genfit" in module:
+            x = sys.modules['__main__']
+            if hasattr(x, 'run'):
+                del x.run
+                print(
+                    "make suicid, until better solution found to ROOT/genfit interference")
+                for f in ROOT.gROOT.GetListOfFiles():
+                    if f.IsWritable() and f.IsOpen(): f.Close()
+                os.system('kill '+str(os.getpid()))
+            if hasattr(x,'fMan'): del x.fMan
+            if hasattr(x,'fRun'): del x.fRun
+            return
+    print("Genfit is not used. Exit normally")
 
 def configure(darkphoton=None):
    ROOT.gROOT.ProcessLine('#include "'+os.environ["FAIRSHIP"]+'/shipdata/ShipGlobals.h"')
